monitoring/requirements.txt
# Python dependencies for monitoring scripts
psycopg2-binary==2.9.9
django==4.2.7
pandas==2.1.3
matplotlib==3.8.2
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
tabulate==0.9.0
